+++ plugins/content/kunenadiscuss/language/en-GB/en-GB.plg_content_kunenadiscuss.ini
@@ -106,3 +106,7 @@ PLG_KUNENADISCUSS_SHOW_LOGIN_PUBLIC_USERS_DESC="Enabling this will show the comm
 PLG_KUNENADISCUSS_GEN_GUEST="<strong>INFO</strong>: You are posting the message as a 'Guest'"
 PLG_KUNENADISCUSS_SHOW_ARTICLE="Show on Articles"
 PLG_KUNENADISCUSS_SHOW_ARTICLE_DESC="What to display on article view pages"
+
+;Discuss 6.0
+PLG_KUNENADISCUSS_LAYOUT="Layout"
+PLG_KUNENADISCUSS_LAYOUT_DESC="Select the Layout to use."
