+++ plugins/content/kunenadiscuss/media/css/aurelia.css
@@ -8,7 +8,12 @@
  * @link https://www.kunena.org
  **/
 
-.kunenadiscuss {
+.kunenadiscuss .avatar > img {
+	 width: 100%;
+	 object-fit: cover;
+ }
+
+/* .kunenadiscuss {
 	border-bottom:1px solid #f4f4f4;
 	padding-bottom: 10px;
 	margin-bottom: 10px;
@@ -262,4 +267,4 @@ div.kdiscuss-content-cover .code {
 .glyphicon.glyphicon-one-fine-dot:before {
 	content: "\25cf";
 	font-size: 1.5em;
-}
+} */
