+++ plugins/content/kunenadiscuss/media/css/default.css
@@ -1,0 +1,270 @@
+@charset "UTF-8";
+/**
+ * Kunena Discuss Plugin
+ * @package Kunena.plg_content_kunenadiscuss
+ *
+ * @copyright (C) 2008 - 2019 Kunena Team. All rights reserved.
+ * @license http://www.gnu.org/copyleft/gpl.html GNU/GPL
+ * @link https://www.kunena.org
+ **/
+
+.kunenadiscuss .avatar > img {
+	 width: 100%;
+	 object-fit: cover;
+ }
+
+/* .kunenadiscuss {
+	border-bottom:1px solid #f4f4f4;
+	padding-bottom: 10px;
+	margin-bottom: 10px;
+}
+
+.kdiscuss {
+	border: 1px solid #f4f4f4;
+	padding: 0 5px 5px;
+}
+.kdiscuss-form table,
+.kdiscuss-form tr,
+.kdiscuss-form td {
+	border: 0px;
+}
+.kdiscuss-reply-header {
+	padding: 5px;
+	font-size: 12px;
+	background: #f4f4f4;
+}
+.kdiscuss-reply-body {
+	overflow: hidden;
+	border-bottom: 1px solid #f4f4f4;
+}
+.kdiscuss-id {
+	float: right;
+	font-size: 11px;
+}
+.kdiscuss-form .kdiscuss-title {
+	margin-bottom: 5px;
+}
+.kdiscuss-item {
+	clear: both;
+	display: block;
+}
+.kdiscuss-item1,
+.kdiscuss-item2 {
+	color: #666;
+	background: #ffffff;
+}
+.kdiscuss-avatar a:link,
+.kdiscuss-avatar a:visited,
+.kdiscuss-avatar a:hover,
+.kdiscuss-avatar a:active,
+.kdiscuss-avatar a:focus {
+	background: inherit !important;
+}
+.kdiscuss-avatar {
+	float: left;
+	width: 48px;
+	margin-right: 15px;
+}
+.kdiscuss-avatar img {
+	margin: 10px 15px 10px 0px;
+	max-width: 48px !important;
+	border: 0px;
+}
+.kdiscuss-content-cover{
+	padding: 5px 10px 20px 0px;
+	overflow: auto;
+	width: 90%;
+}
+.kdiscuss-itemid-cover {
+	padding: 15px;
+}
+.kdiscuss-title {
+	font-size: 14px;
+	font-weight: bold;
+	padding: 5px;
+	text-align: left;
+	margin: 10px 0px 0px;
+	clear: both;
+	background: #d4dee3;
+	border-bottom: 1px solid #c3d9e1;
+}
+.kdiscuss-subject {
+	font-family: Arial, Helvetica, sans-serif;
+	font-size: 12px;
+	font-weight: bold;
+	color: #E86D29;
+}
+.kdiscuss-username a {
+	font-family: Arial, Helvetica, sans-serif;
+	font-size: 12px;
+	font-weight: bold;
+}
+.kdiscuss-date {
+	font-family: Arial, Helvetica, sans-serif;
+	font-size: 11px;
+	font-weight: normal;
+	color: #666;
+	padding-left: 20px;
+	padding-top: 2px;
+}
+.kdiscuss-username,
+.kdiscuss-username a {
+	font-size: 11px;
+}
+.kdiscuss-text {
+	font-size: 11px;
+	padding: 5px 0;
+	float: left;
+	width: 70%;
+}
+.kdiscuss-text img {
+	max-width: 450px;
+}
+.kdiscuss-more {
+	margin: 10px 5px;
+	font-size: 11px;
+}
+div.kdiscuss-content-cover pre,
+div.kdiscuss-content-cover .code {
+	border-color: #CCCCCC rgb(204, 204, 204) rgb(204, 204, 204) rgb(244, 169, 79);
+	border-style: solid;
+	border-width: 1px 1px 1px 5px;
+	font-family: "Courier New",monospace;
+	font-size: 100%;
+	font-size-adjust: none;
+	font-stretch: normal;
+	font-style: normal;
+	font-variant: normal;
+	font-weight: normal;
+	line-height: 1.5;
+	margin: 5px 0pt 15px;
+	padding: 10px 15px;
+	overflow: scroll;
+	width: 100%;
+}
+.kdiscuss-quick-post-label {
+	font-weight: bold;
+}
+.kdiscuss-fatal {
+	font-weight: bold;
+	color: red;
+}
+#Kunena div.kmsgtext-article, #Kunena div.kmsgtext-quote, #Kunena div.kmsgtext-hide, #Kunena div.kmsgtext-confidential {
+	display: block;
+	border: 1px dotted;
+	margin: 5px 0;
+	padding: 5px;
+	font-style: italic;
+}
+#Kunena div.kmsgtext-article, #Kunena div.kmsgtext-confidential {
+	border-color: #BFC3C6;
+	font-family: Arial,Helvetica,sans-serif;
+	background-color: #F2F1EE;
+}
+.ktext {
+	width: 100%;
+}
+.kmsgtext-quote, .kmsgtext-hide {
+	background-color: #f4f4f4;
+	border-color: #BFC3C6;
+	font-family: Arial,Helvetica,sans-serif;
+	border-radius: 5px;
+	padding: 5px;
+}
+
+.kdiscuss-title.login-discuss {
+	margin-bottom: 40px;
+}
+.klogin-to-discuss {
+	background-color: #ffffff;
+	border: 1px solid #d5d5d5;
+	border-radius: 3px;
+	box-shadow: none;
+	box-sizing: border-box;
+	color: #6e6e6e;
+	padding: 12px 15px;
+}
+.klogin-to-discuss:hover {
+	color: #E35614;
+	border: 1px solid #E35614;
+}
+.kpublic {
+	border-bottom: 0 none !important;
+}
+
+.img-circle {
+	margin-top: 0.5em;
+	position: relative;
+
+	-webkit-border-radius: 50%;
+	-moz-border-radius: 50%;
+	border-radius: 50%;
+
+	-webkit-box-shadow: 0 0 0 3px #fff, 0 0 0 4px #ccc, 0 2px 5px 4px rgba(0, 0, 0, 0.1);;
+	-moz-box-shadow: 0 0 0 3px #fff, 0 0 0 4px #ccc, 0 2px 5px 4px rgba(0, 0, 0, 0.1);
+	box-shadow: 0 0 0 3px #fff, 0 0 0 4px #ccc, 0 2px 5px 4px rgba(0, 0, 0, 0.1);
+}
+
+.kwho-guest {
+	color: #666666;
+}
+
+.kwho-user {
+	color: #5388B4;
+}
+
+.kwho-moderator {
+	color: #0000FF;
+}
+
+.kwho-globalmod,
+.kwho-globalmoderator {
+	color: #800000;
+}
+
+.kwho-admin {
+	color: #FF0000;
+}
+
+.kwho-blocked,
+.kwho-banned {
+	color: #A39D49;
+}
+
+.b2-3x {
+	padding-top: 10px;
+	font-size: 32px;
+}
+.b2-7x {
+	padding-top: 10px;
+	font-size: 52px;
+}
+.user-circle.user-default {
+	width: 32px;
+	height: 32px;
+	text-align: center;
+	font-size: 24px;
+	line-height: 1.33;
+	border-radius: 50%;
+	border: 1px solid;
+}
+.user-circle.user-xl {
+	width: 70px;
+	height: 70px;
+	padding: 10px 16px;
+	font-size: 52px;
+	line-height: 1.33;
+	border-radius: 50%;
+	border: 1px solid;
+}
+.icon.icon-one-fine-dot {
+	margin-right: 0;
+}
+.icon.icon-one-fine-dot:before {
+	content: "\25cf";
+	font-size: 2em;
+}
+.glyphicon.glyphicon-one-fine-dot:before {
+	content: "\25cf";
+	font-size: 1.5em;
+} */
