+++ plugins/content/kunenadiscuss/tmpl/aurelia_bottom.php
@@ -1,0 +1,14 @@
+<?php
+
+/**
+ * Kunena Discuss Plugin
+ *
+ * @package       Kunena.plg_content_kunenadiscuss
+ *
+ * @copyright (C) 2008 - 2019 Kunena Team. All rights reserved.
+ * @license       http://www.gnu.org/copyleft/gpl.html GNU/GPL
+ * @link          https://www.kunena.org
+ **/
+defined('_JEXEC') or die('');
+
+include __DIR__ . '/aurelia_left.php';
